Moves druid version zip classes into own namespace.
closes #2065
